Support ECDSA with and without recovery key
