+++ src/test/java/sign/ECDSASecp256k1Test.java
@@ -2,20 +2,14 @@ package sign;
 
 import com.ltonetwork.seasalt.Binary;
 import com.ltonetwork.seasalt.KeyPair;
-import com.ltonetwork.seasalt.hash.Hasher;
 import com.ltonetwork.seasalt.sign.ECDSA;
 import com.ltonetwork.seasalt.sign.ECDSASignature;
+import org.bouncycastle.asn1.sec.SECNamedCurves;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
-import org.web3j.crypto.ECKeyPair;
-import org.web3j.crypto.Hash;
-import org.web3j.crypto.Keys;
-import org.web3j.crypto.Sign;
 
-import java.math.BigInteger;
 import java.nio.charset.StandardCharsets;
-import java.security.*;
 import java.util.Random;
 
 public class ECDSASecp256k1Test {
@@ -24,7 +18,7 @@ public class ECDSASecp256k1Test {
 
     @BeforeEach
     public void init() {
-        secp256k1 = new ECDSA("secp256k1");
+        secp256k1 = new ECDSA(SECNamedCurves.getByName("secp256k1"));
     }
 
     @Test
@@ -59,94 +53,51 @@ public class ECDSASecp256k1Test {
 
     @Test
     public void testSigns() {
-        byte[] b = new byte[20];
-        new Random().nextBytes(b);
-        byte[] msgHash = new Hasher("Keccak-256").hash(b).getBytes();
-
         KeyPair kp = secp256k1.keyPair();
+        byte[] msg = "test".getBytes(StandardCharsets.UTF_8);
 
         Assertions.assertDoesNotThrow(() -> {
-            secp256k1.signDetached(msgHash, kp);
+            secp256k1.signDetached(msg, kp);
         });
     }
 
     @Test
     public void testVerify() {
-        byte[] b = new byte[20];
-        new Random().nextBytes(b);
-        byte[] msgHash = new Hasher("Keccak-256").hash(b).getBytes();
-
         KeyPair kp = secp256k1.keyPair();
+        byte[] msg = "test".getBytes(StandardCharsets.UTF_8);
+        Binary sig = secp256k1.signDetached(msg, kp.getPrivateKey().getBytes());
+        ECDSASignature sig2 = secp256k1.signDetached(msg, kp.getPrivateKey().getBytes());
 
-        ECDSASignature sig = secp256k1.signDetached(msgHash, kp.getPrivateKey().getBytes());
-
-        Assertions.assertTrue(secp256k1.verify(msgHash, sig, kp));
+        Assertions.assertTrue(secp256k1.verify(msg, sig, kp.getPublicKey().getBytes()));
+        Assertions.assertTrue(secp256k1.verify(msg, sig2, kp.getPublicKey().getBytes()));
     }
+
+    //FIXME:
+//    @Test
+//    public void testVerifyDifferentKp() {
+//        ECDSARecovery secp256k1Recovery = new ECDSARecovery(SECNamedCurves.getByName("secp256k1"));
+//
+//        KeyPair kpRecovery = secp256k1Recovery.keyPair();
+//
+//        KeyPair kp = secp256k1.keyPair();
+//
+//        System.out.println(kpRecovery.getPrivateKey().getBytes().length);
+//        System.out.println(kp.getPrivateKey().getBytes().length);
+//        System.out.println(kpRecovery.getPublicKey().getBytes().length);
+//        System.out.println(kp.getPublicKey().getBytes().length);
+//
+//        byte[] msg = "test".getBytes(StandardCharsets.UTF_8);
+//        ECDSASignature sig = secp256k1Recovery.signDetached(msg, kp.getPrivateKey().getBytes());
+//
+//        Assertions.assertTrue(secp256k1Recovery.verify(msg, sig, kp.getPublicKey().getBytes()));
+//    }
 
     @Test
     public void testVerifyFail() {
-        byte[] b = new byte[20];
-        new Random().nextBytes(b);
-        byte[] msgHash = new Hasher("Keccak-256").hash(b).getBytes();
-
         KeyPair kp = secp256k1.keyPair();
-        Binary sig = secp256k1.signDetached(msgHash, kp.getPrivateKey().getBytes());
+        byte[] msg = "test".getBytes(StandardCharsets.UTF_8);
+        Binary sig = secp256k1.signDetached(msg, kp.getPrivateKey().getBytes());
 
         Assertions.assertFalse(secp256k1.verify("fail".getBytes(StandardCharsets.UTF_8), sig, kp));
-    }
-
-    @Test
-    public void testVerifyWithWeb3() throws SignatureException {
-        byte[] b = new byte[20];
-        new Random().nextBytes(b);
-        byte[] msgHash = new Hasher("Keccak-256").hash(b).getBytes();
-
-        KeyPair kpSeaSalt = secp256k1.keyPair();
-        ECDSASignature sigSeaSalt = secp256k1.signDetached(msgHash, kpSeaSalt.getPrivateKey().getBytes());
-
-        Sign.SignatureData sigWeb3 = new Sign.SignatureData(sigSeaSalt.getV(), sigSeaSalt.getR(), sigSeaSalt.getS());
-        BigInteger recoveredWeb3 = Sign.signedMessageHashToKey(msgHash, sigWeb3);
-        Assertions.assertArrayEquals(kpSeaSalt.getPublicKey().getBytes(), recoveredWeb3.toByteArray());
-    }
-
-    @Test
-    public void testSignWithWeb3() throws InvalidAlgorithmParameterException, NoSuchAlgorithmException, NoSuchProviderException {
-        byte[] b = new byte[20];
-        new Random().nextBytes(b);
-        byte[] msgHash = Hash.sha3(b);
-
-        ECKeyPair kpWeb3 = Keys.createEcKeyPair();
-        Sign.SignatureData sigWeb3 = Sign.signMessage(msgHash, kpWeb3, false);
-
-        ECDSASignature sigSeaSalt = new ECDSASignature(sigWeb3.getR(), sigWeb3.getS(), sigWeb3.getV());
-
-        Assertions.assertTrue(secp256k1.verify(msgHash, sigSeaSalt, kpWeb3.getPublicKey().toByteArray()));
-    }
-
-    @Test
-    public void testWeb3SeasaltFull() throws InvalidAlgorithmParameterException, NoSuchAlgorithmException, NoSuchProviderException, SignatureException {
-        byte[] b = new byte[20];
-        new Random().nextBytes(b);
-        byte[] msgHash = Hash.sha3(b);
-
-        // WEB3J
-        ECKeyPair kpWeb3 = Keys.createEcKeyPair();
-        Sign.SignatureData sigWeb3 = Sign.signMessage(msgHash, kpWeb3, false);
-
-        BigInteger recoveredWeb3 = Sign.signedMessageHashToKey(msgHash, sigWeb3);
-
-        Assertions.assertEquals(kpWeb3.getPublicKey(), recoveredWeb3);
-
-        // SEASALT
-        KeyPair kpSeaSalt = secp256k1.keyPairFromSecretKey(kpWeb3.getPrivateKey().toByteArray());
-        ECDSASignature sigSeaSalt = secp256k1.signDetached(msgHash, kpSeaSalt.getPrivateKey().getBytes());
-
-        Sign.SignatureData sig2Web3 = new Sign.SignatureData(sigSeaSalt.getV(), sigSeaSalt.getR(), sigSeaSalt.getS());
-        BigInteger recoveredSeaSalt = Sign.signedMessageHashToKey(msgHash, sig2Web3);
-
-        Assertions.assertArrayEquals(kpSeaSalt.getPublicKey().getBytes(), recoveredSeaSalt.toByteArray());
-
-        Assertions.assertEquals(recoveredWeb3, recoveredSeaSalt);
-
     }
 }
